fix: JS function arguments determination & refactor JSObject code (#14871)
* fix: JsObject parsing

* fix jest test and add types

* remove logs from test

* Add jest test to verify this

* Add perf metric and

* fix: jsObject stale value
- refactor updateJSCollectionInDataTree

* Refactor JSObject code

* fix import path

* refactor and fix

* undo debugging change

* Add comments

* Add Ast for getting function params

* Fix and add test

* remove comment

* remove comment

* Resolve comments
