+++ app/client/src/workers/DataTreeEvaluator/mockData/NestedArrayAccessorTree.ts
