+++ app/client/src/workers/evaluationUtils.ts
@@ -18,15 +18,12 @@ import {
   DataTreeWidget,
   ENTITY_TYPE,
   DataTreeJSAction,
-  EvaluationSubstitutionType,
   PrivateWidgets,
 } from "entities/DataTree/dataTreeFactory";
 import _ from "lodash";
 import { WidgetTypeConfigMap } from "utils/WidgetFactory";
 import { ValidationConfig } from "constants/PropertyControlConstants";
 import { Severity } from "entities/AppsmithConsole";
-import { ParsedBody, ParsedJSSubAction } from "utils/JSPaneUtils";
-import { Variable } from "entities/JSCollection";
 import { PluginType } from "entities/Action";
 import { klona } from "klona/full";
 import { warn as logWarn } from "loglevel";
@@ -638,28 +635,6 @@ export function getSafeToRenderDataTree(
   }, tree);
 }
 
-export const STRIP_COMMENTS = /((\/\/.*$)|(\/\*[\s\S]*?\*\/))/gm;
-export const ARGUMENT_NAMES = /([^\s,]+)/g;
-
-export function getParams(func: any) {
-  const fnStr = func.toString().replace(STRIP_COMMENTS, "");
-  const args: Array<Variable> = [];
-  let result = fnStr
-    .slice(fnStr.indexOf("(") + 1, fnStr.indexOf(")"))
-    .match(ARGUMENT_NAMES);
-  if (result === null) result = [];
-  if (result && result.length) {
-    result.forEach((arg: string) => {
-      const element = arg.split("=");
-      args.push({
-        name: element[0],
-        value: element[1],
-      });
-    });
-  }
-  return args;
-}
-
 export const addErrorToEntityProperty = (
   errors: EvaluationError[],
   dataTree: DataTree,
@@ -708,254 +683,6 @@ export const isDynamicLeaf = (unEvalTree: DataTree, propertyPath: string) => {
     relativePropertyPath in entity.reactivePaths ||
     (isWidget(entity) && relativePropertyPath in entity.triggerPaths)
   );
-};
-
-export const updateJSCollectionInDataTree = (
-  parsedBody: ParsedBody,
-  jsCollection: DataTreeJSAction,
-  dataTree: DataTree,
-) => {
-  const modifiedDataTree: any = dataTree;
-  const functionsList: Array<string> = [];
-  const varList: Array<string> = jsCollection.variables;
-  Object.keys(jsCollection.meta).forEach((action) => {
-    functionsList.push(action);
-  });
-
-  if (parsedBody.actions && parsedBody.actions.length > 0) {
-    for (let i = 0; i < parsedBody.actions.length; i++) {
-      const action = parsedBody.actions[i];
-      if (jsCollection.hasOwnProperty(action.name)) {
-        if (jsCollection[action.name] !== action.body) {
-          const data = _.get(
-            modifiedDataTree,
-            `${jsCollection.name}.${action.name}.data`,
-            {},
-          );
-          _.set(
-            modifiedDataTree,
-            `${jsCollection.name}.${action.name}`,
-            new String(action.body),
-          );
-
-          _.set(
-            modifiedDataTree,
-            `${jsCollection.name}.${action.name}.data`,
-            data,
-          );
-        }
-      } else {
-        const reactivePaths = jsCollection.reactivePaths;
-        reactivePaths[action.name] =
-          EvaluationSubstitutionType.SMART_SUBSTITUTE;
-        reactivePaths[`${action.name}.data`] =
-          EvaluationSubstitutionType.TEMPLATE;
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.reactivePaths`,
-          reactivePaths,
-        );
-        const dynamicBindingPathList = jsCollection.dynamicBindingPathList;
-        dynamicBindingPathList.push({ key: action.name });
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.dynamicBindingPathList`,
-          dynamicBindingPathList,
-        );
-        const dependencyMap = jsCollection.dependencyMap;
-        dependencyMap["body"].push(action.name);
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.dependencyMap`,
-          dependencyMap,
-        );
-        const meta = jsCollection.meta;
-        meta[action.name] = {
-          arguments: action.arguments,
-          isAsync: false,
-          confirmBeforeExecute: false,
-        };
-        _.set(modifiedDataTree, `${jsCollection.name}.meta`, meta);
-        const data = _.get(
-          modifiedDataTree,
-          `${jsCollection.name}.${action.name}.data`,
-          {},
-        );
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.${action.name}`,
-          new String(action.body.toString()),
-        );
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.${action.name}.data`,
-          data,
-        );
-      }
-    }
-  }
-  if (functionsList && functionsList.length > 0) {
-    for (let i = 0; i < functionsList.length; i++) {
-      const preAction = functionsList[i];
-      const existed = parsedBody.actions.find(
-        (js: ParsedJSSubAction) => js.name === preAction,
-      );
-      if (!existed) {
-        const reactivePaths = jsCollection.reactivePaths;
-        delete reactivePaths[preAction];
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.reactivePaths`,
-          reactivePaths,
-        );
-        let dynamicBindingPathList = jsCollection.dynamicBindingPathList;
-        dynamicBindingPathList = dynamicBindingPathList.filter(
-          (path) => path["key"] !== preAction,
-        );
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.dynamicBindingPathList`,
-          dynamicBindingPathList,
-        );
-        const dependencyMap = jsCollection.dependencyMap["body"];
-        const removeIndex = dependencyMap.indexOf(preAction);
-        if (removeIndex > -1) {
-          const updatedDMap = dependencyMap.filter(
-            (item) => item !== preAction,
-          );
-          _.set(
-            modifiedDataTree,
-            `${jsCollection.name}.dependencyMap.body`,
-            updatedDMap,
-          );
-        }
-        const meta = jsCollection.meta;
-        delete meta[preAction];
-        _.set(modifiedDataTree, `${jsCollection.name}.meta`, meta);
-        delete modifiedDataTree[`${jsCollection.name}`][`${preAction}`];
-        delete modifiedDataTree[`${jsCollection.name}`][`${preAction}.data`];
-      }
-    }
-  }
-  if (parsedBody.variables.length) {
-    for (let i = 0; i < parsedBody.variables.length; i++) {
-      const newVar = parsedBody.variables[i];
-      const existedVar = varList.indexOf(newVar.name);
-      if (existedVar > -1) {
-        const existedVarVal = jsCollection[newVar.name];
-        if (
-          (!!existedVarVal && existedVarVal.toString()) !==
-            (newVar.value && newVar.value.toString()) ||
-          (!existedVarVal && !!newVar)
-        ) {
-          _.set(
-            modifiedDataTree,
-            `${jsCollection.name}.${newVar.name}`,
-            newVar.value,
-          );
-        }
-      } else {
-        varList.push(newVar.name);
-        const reactivePaths = jsCollection.reactivePaths;
-        reactivePaths[newVar.name] =
-          EvaluationSubstitutionType.SMART_SUBSTITUTE;
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.reactivePaths`,
-          reactivePaths,
-        );
-        const dynamicBindingPathList = jsCollection.dynamicBindingPathList;
-        dynamicBindingPathList.push({ key: newVar.name });
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.dynamicBindingPathList`,
-          dynamicBindingPathList,
-        );
-
-        _.set(modifiedDataTree, `${jsCollection.name}.variables`, varList);
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.${newVar.name}`,
-          newVar.value,
-        );
-      }
-    }
-    let newVarList: Array<string> = varList;
-    for (let i = 0; i < varList.length; i++) {
-      const varListItem = varList[i];
-      const existsInParsed = parsedBody.variables.find(
-        (item) => item.name === varListItem,
-      );
-      if (!existsInParsed) {
-        const reactivePaths = jsCollection.reactivePaths;
-        delete reactivePaths[varListItem];
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.reactivePaths`,
-          reactivePaths,
-        );
-
-        let dynamicBindingPathList = jsCollection.dynamicBindingPathList;
-        dynamicBindingPathList = dynamicBindingPathList.filter(
-          (path) => path["key"] !== varListItem,
-        );
-        _.set(
-          modifiedDataTree,
-          `${jsCollection.name}.dynamicBindingPathList`,
-          dynamicBindingPathList,
-        );
-
-        newVarList = newVarList.filter((item) => item !== varListItem);
-        delete modifiedDataTree[`${jsCollection.name}`][`${varListItem}`];
-      }
-    }
-    if (newVarList.length) {
-      _.set(modifiedDataTree, `${jsCollection.name}.variables`, newVarList);
-    }
-  }
-  return modifiedDataTree;
-};
-
-export const removeFunctionsAndVariableJSCollection = (
-  dataTree: DataTree,
-  entity: DataTreeJSAction,
-) => {
-  const modifiedDataTree: any = dataTree;
-  const functionsList: Array<string> = [];
-  Object.keys(entity.meta).forEach((action) => {
-    functionsList.push(action);
-  });
-  //removed variables
-  const varList: Array<string> = entity.variables;
-  _.set(modifiedDataTree, `${entity.name}.variables`, []);
-  for (let i = 0; i < varList.length; i++) {
-    const varName = varList[i];
-    delete modifiedDataTree[`${entity.name}`][`${varName}`];
-  }
-  //remove functions
-  let dynamicBindingPathList = entity.dynamicBindingPathList;
-  const reactivePaths = entity.reactivePaths;
-  const meta = entity.meta;
-  let dependencyMap = entity.dependencyMap["body"];
-  for (let i = 0; i < functionsList.length; i++) {
-    const actionName = functionsList[i];
-    delete reactivePaths[actionName];
-    delete meta[actionName];
-    delete modifiedDataTree[`${entity.name}`][`${actionName}`];
-    dynamicBindingPathList = dynamicBindingPathList.filter(
-      (path: any) => path["key"] !== actionName,
-    );
-    dependencyMap = dependencyMap.filter((item: any) => item !== actionName);
-  }
-  _.set(modifiedDataTree, `${entity.name}.reactivePaths`, reactivePaths);
-  _.set(
-    modifiedDataTree,
-    `${entity.name}.dynamicBindingPathList`,
-    dynamicBindingPathList,
-  );
-  _.set(modifiedDataTree, `${entity.name}.dependencyMap.body`, dependencyMap);
-  _.set(modifiedDataTree, `${entity.name}.meta`, meta);
-  return modifiedDataTree;
 };
 
 export const addWidgetPropertyDependencies = ({
